+++ importlib_resources/trees.py
@@ -6,13 +6,29 @@ import contextlib
 
 from ._compat import (
     Path, package_spec, FileNotFoundError, ZipPath,
-    singledispatch,
+    singledispatch, suppress,
     )
 
 
 def from_package(package):
-    """Return a Traversable object for the given package"""
+    """
+    Return a Traversable object for the given package.
+
+    """
     spec = package_spec(package)
+    return from_traversable_resources(spec) or fallback_resources(spec)
+
+
+def from_traversable_resources(spec):
+    """
+    If the spec.loader implements TraversableResources,
+    directly or implicitly, it will have a ``files()`` method.
+    """
+    with suppress(AttributeError):
+        return spec.loader.files()
+
+
+def fallback_resources(spec):
     package_directory = Path(spec.origin).parent
     try:
         archive_path = spec.loader.archive
@@ -44,7 +60,7 @@ def _tempfile(reader):
 @contextlib.contextmanager
 def as_file(path):
     """
-    Given a path-like object, return that object as a
+    Given a Traversable object, return that object as a
     path on the local file system in a context manager.
     """
     with _tempfile(path.read_bytes) as local:
@@ -55,6 +71,6 @@ def as_file(path):
 @contextlib.contextmanager
 def _(path):
     """
-    Degenerate behavior for pathlib.Path objects
+    Degenerate behavior for pathlib.Path objects.
     """
     yield path